requirements.txt
boto==2.43.0
dj-database-url==0.4.1
Django==1.9.1
django-cors-headers==1.1.0
django-filter==1.0.1
django-pgfields==1.4.4
django-storages==1.5.1
djangorestframework==3.2.4
drfdocs==0.0.11
gunicorn==19.6.0
hashids==1.1.0
oauthlib==2.0.0
Pillow==2.8.2
psycopg2==2.6.1
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.10
pytz==2016.7
qrcode==5.3
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
whitenoise==3.2.2
